Ajout ennemies: bat

+++ src/Systems/Commons/Systems-Camera.hpp
@@ -15,9 +15,9 @@ void cameraSys(World& world) {
         const auto& windowSize = vulkanEngine.window.getSize();
 
         if (160.f * (static_cast<float>(windowSize.x) / static_cast<float>(windowSize.y)) > 160.f) {
-            inGameView.setSize(160.f * (static_cast<float>(windowSize.x) / static_cast<float>(windowSize.y)), 144.f);
+            inGameView.setSize(static_cast<int>(160.f * (static_cast<float>(windowSize.x) / static_cast<float>(windowSize.y))), 144);
         } else {
-            inGameView.setSize(160.f, 144.f * (static_cast<float>(windowSize.y) / static_cast<float>(windowSize.x)));
+            inGameView.setSize(160.f, static_cast<int>(144.f * (static_cast<float>(windowSize.y) / static_cast<float>(windowSize.x))));
         }
         inGameView.setCenter(trans.getPosition());
     }
